+++ frontend/src/components/PopUp/SubmitSuccessPopUp.styled.ts
